+++ module/settings/settings-helpers.mjs
@@ -239,6 +239,17 @@ export default class SettingsHelpers {
             requiresReload: false,
         });
 
+        game.settings.register(module, "ShowAllConditionalDefenses", {
+            name: game.i18n.localize("Settings.ShowAllConditionalDefenses.Name"),
+            hint: game.i18n.localize("Settings.ShowAllConditionalDefenses.Hint"),
+            scope: "world",
+            config: true,
+            type: Boolean,
+            default: false,
+            //onChange: () => ui.combat.render(),
+            requiresReload: false,
+        });
+
         game.settings.register(module, "alphaTesting", {
             name: game.i18n.localize("Settings.AlphaTesting.Name"),
             hint: game.i18n.localize("Settings.AlphaTesting.Hint"),